Revert "add login page"
This reverts commit 8fb203529dace045d2a43263855c5dbbdb80dd6c.
